+++ scripts/site-wizard/ao3_site_wizard.js
@@ -5,7 +5,7 @@
 // @author        BlackBatcat
 // @match         *://archiveofourown.org/*
 // @license       MIT
-// @require       https://update.greasyfork.org/scripts/552743/1850777/AO3%3A%20Menu%20Helpers%20Library.js?v=2.2.3
+// @require       https://update.greasyfork.org/scripts/552743/1853381/AO3%3A%20Menu%20Helpers%20Library.js?v=2.2.3
 // @grant         none
 // @run-at        document-start
 // ==/UserScript==