+++ scripts/custom-favorites/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
 
-## Unreleased
+## v1.0.0
 - 
+- **Updated to Menu Helpers Library v2.2.3** — Fixed menu bug on AO3 homepage.